fix: 景點明細車程數字格式對齊行程表（單一 formatDistance SoT）
EditEntryPage「從上一站移動」與車程 dialog 的距離格式與 timeline TravelPill
drift：≥10km 時 timeline 顯「30 km」但景點明細顯「30.0 km」。三份重複 formatter
（TravelPill / TravelPillDialog / EditEntryPage formatKm）整併成單一 SoT
formatDistance（src/lib/timelineUtils.ts），並修 EditEntryPage crossRegionWarning
第 4 份 inline formatter。+11 unit tests 鎖格式，永久消除 drift。
